requirements.txt
Pillow>=10.0
wandb>=0.16
